Factor out parseSse into another file
